requirements.txt
asgiref==3.8.1
Django==5.0.6
pillow==10.4.0
sqlparse==0.5.0
typing_extensions==4.12.2
django-froala-editor==4.0.13
djangorestframework==3.13.1